# ddl: skip delay for async commit during merging temp index